am243x: eclipse-threadx: Fix various MISRA warnings.
- Fix for various MISRA warnings in the ThreadX specific source code.

Fixes: SITSW-5614

+++ source/kernel/threadx/dpl/common/ClockP_threadx.c
@@ -157,7 +157,7 @@ uint32_t ClockP_isActive(ClockP_Object *handle)
         active = TX_FALSE;
     }
 
-    return active;
+    return (uint32_t)active;
 }
 
 void ClockP_Params_init(ClockP_Params *params)
@@ -209,7 +209,7 @@ void ClockP_usleep(uint32_t usec)
     endTime = curTime + usec;
 
     if (usec >= gClockCtrl.usecPerTick) {
-        ticksToSleep = usec / gClockCtrl.usecPerTick;
+        ticksToSleep = (uint32_t)(usec / gClockCtrl.usecPerTick);
         ClockP_sleepTicks(ticksToSleep);
     }
     else
